requirements2.txt
--index-url=https://download.pytorch.org/whl/cu128
torchcodec==0.8.1
